move to hpp/cpp structure, create conversions file
