+++ depth_image_proc/src/point_cloud_xyz_radial.cpp
@@ -29,98 +29,22 @@
 // LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
+#include "depth_image_proc/point_cloud_xyz_radial.hpp"
 #include <rclcpp/rclcpp.hpp>
 #include <image_transport/image_transport.hpp>
 #include <sensor_msgs/image_encodings.hpp>
 #include <image_geometry/pinhole_camera_model.h>
 #include <depth_image_proc/depth_traits.hpp>
-#include <depth_image_proc/depth_conversions.hpp>
+#include <depth_image_proc/conversions.hpp>
 #include <depth_image_proc/visibility.h>
 #include <sensor_msgs/point_cloud2_iterator.hpp>
+
 #include <memory>
 #include <vector>
 #include <limits>
 
 namespace depth_image_proc
 {
-
-namespace enc = sensor_msgs::image_encodings;
-
-class PointCloudXyzRadialNode : public rclcpp::Node
-{
-public:
-  DEPTH_IMAGE_PROC_PUBLIC PointCloudXyzRadialNode(const rclcpp::NodeOptions & options);
-
-private:
-  // Subscriptions
-  image_transport::CameraSubscriber sub_depth_;
-  int queue_size_;
-
-  // Publications
-  std::mutex connect_mutex_;
-  using PointCloud = sensor_msgs::msg::PointCloud2;
-  rclcpp::Publisher<PointCloud>::SharedPtr pub_point_cloud_;
-
-  std::vector<double> D_;
-  std::array<double, 9> K_;
-
-  uint32_t width_;
-  uint32_t height_;
-
-  cv::Mat binned;
-
-  void connectCb();
-
-  void depthCb(
-    const sensor_msgs::msg::Image::ConstSharedPtr & depth_msg,
-    const sensor_msgs::msg::CameraInfo::ConstSharedPtr & info_msg);
-
-  // Handles float or uint16 depths
-  template<typename T>
-  void convert(
-    const sensor_msgs::msg::Image::ConstSharedPtr & depth_msg,
-    PointCloud::SharedPtr & cloud_msg);
-
-  rclcpp::Logger logger_ = rclcpp::get_logger("PointCloudXyzRadialNode");
-};
-
-cv::Mat initMatrix(cv::Mat cameraMatrix, cv::Mat distCoeffs, int width, int height, bool radial)
-{
-  int i, j;
-  int totalsize = width * height;
-  cv::Mat pixelVectors(1, totalsize, CV_32FC3);
-  cv::Mat dst(1, totalsize, CV_32FC3);
-
-  cv::Mat sensorPoints(cv::Size(height, width), CV_32FC2);
-  cv::Mat undistortedSensorPoints(1, totalsize, CV_32FC2);
-
-  std::vector<cv::Mat> ch;
-  for (j = 0; j < height; j++) {
-    for (i = 0; i < width; i++) {
-      cv::Vec2f & p = sensorPoints.at<cv::Vec2f>(i, j);
-      p[0] = i;
-      p[1] = j;
-    }
-  }
-
-  sensorPoints = sensorPoints.reshape(2, 1);
-
-  cv::undistortPoints(sensorPoints, undistortedSensorPoints, cameraMatrix, distCoeffs);
-
-  ch.push_back(undistortedSensorPoints);
-  ch.push_back(cv::Mat::ones(1, totalsize, CV_32FC1));
-  cv::merge(ch, pixelVectors);
-
-  if (radial) {
-    for (i = 0; i < totalsize; i++) {
-      normalize(
-        pixelVectors.at<cv::Vec3f>(i),
-        dst.at<cv::Vec3f>(i));
-    }
-    pixelVectors = dst;
-  }
-  return pixelVectors.reshape(3, width);
-}
 
 
 PointCloudXyzRadialNode::PointCloudXyzRadialNode(const rclcpp::NodeOptions & options)
@@ -186,50 +110,20 @@ void PointCloudXyzRadialNode::depthCb(
     K_ = info_msg->k;
     width_ = info_msg->width;
     height_ = info_msg->height;
-    binned = initMatrix(cv::Mat_<double>(3, 3, &K_[0]), cv::Mat(D_), width_, height_, true);
+    transform_ = initMatrix(cv::Mat_<double>(3, 3, &K_[0]), cv::Mat(D_), width_, height_, true);
   }
 
+  // Convert Depth Image to Pointcloud
   if (depth_msg->encoding == enc::TYPE_16UC1) {
-    convert<uint16_t>(depth_msg, cloud_msg);
+    convertDepthRadial<uint16_t>(depth_msg, cloud_msg, transform_);
   } else if (depth_msg->encoding == enc::TYPE_32FC1) {
-    convert<float>(depth_msg, cloud_msg);
+    convertDepthRadial<float>(depth_msg, cloud_msg, transform_);
   } else {
     RCLCPP_ERROR(logger_, "Depth image has unsupported encoding [%s]", depth_msg->encoding.c_str());
     return;
   }
 
   pub_point_cloud_->publish(*cloud_msg);
-}
-
-template<typename T>
-void PointCloudXyzRadialNode::convert(
-  const sensor_msgs::msg::Image::ConstSharedPtr & depth_msg,
-  PointCloud::SharedPtr & cloud_msg)
-{
-  // Combine unit conversion (if necessary) with scaling by focal length for computing (X,Y)
-  float bad_point = std::numeric_limits<float>::quiet_NaN();
-
-  sensor_msgs::PointCloud2Iterator<float> iter_x(*cloud_msg, "x");
-  sensor_msgs::PointCloud2Iterator<float> iter_y(*cloud_msg, "y");
-  sensor_msgs::PointCloud2Iterator<float> iter_z(*cloud_msg, "z");
-  const T * depth_row = reinterpret_cast<const T *>(&depth_msg->data[0]);
-  int row_step = depth_msg->step / sizeof(T);
-  for (int v = 0; v < static_cast<int>(cloud_msg->height); ++v, depth_row += row_step) {
-    for (int u = 0; u < static_cast<int>(cloud_msg->width); ++u, ++iter_x, ++iter_y, ++iter_z) {
-      T depth = depth_row[u];
-
-      // Missing points denoted by NaNs
-      if (!DepthTraits<T>::valid(depth)) {
-        *iter_x = *iter_y = *iter_z = bad_point;
-        continue;
-      }
-      const cv::Vec3f & cvPoint = binned.at<cv::Vec3f>(u, v) * DepthTraits<T>::toMeters(depth);
-      // Fill in XYZ
-      *iter_x = cvPoint(0);
-      *iter_y = cvPoint(1);
-      *iter_z = cvPoint(2);
-    }
-  }
 }
 
 }  // namespace depth_image_proc
