+++ depth_image_proc/src/point_cloud_xyzi_radial.cpp
@@ -29,6 +29,7 @@
 // LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
+#include "depth_image_proc/point_cloud_xyzi_radial.hpp"
 #include <rclcpp/rclcpp.hpp>
 #include <image_transport/image_transport.hpp>
 #include <sensor_msgs/image_encodings.hpp>
@@ -38,8 +39,10 @@
 #include <message_filters/sync_policies/exact_time.h>
 #include <image_geometry/pinhole_camera_model.h>
 #include <depth_image_proc/depth_traits.hpp>
+#include <depth_image_proc/conversions.hpp>
 #include <depth_image_proc/visibility.h>
 #include <sensor_msgs/point_cloud2_iterator.hpp>
+
 #include <memory>
 #include <vector>
 #include <string>
@@ -47,108 +50,6 @@
 
 namespace depth_image_proc
 {
-
-using namespace std::placeholders;
-namespace enc = sensor_msgs::image_encodings;
-using SyncPolicy =
-  message_filters::sync_policies::ExactTime<
-  sensor_msgs::msg::Image,
-  sensor_msgs::msg::Image,
-  sensor_msgs::msg::CameraInfo>;
-
-class PointCloudXyziRadialNode : public rclcpp::Node
-{
-public:
-  DEPTH_IMAGE_PROC_PUBLIC PointCloudXyziRadialNode(const rclcpp::NodeOptions & options);
-
-private:
-  using PointCloud = sensor_msgs::msg::PointCloud2;
-  using Image = sensor_msgs::msg::Image;
-  using CameraInfo = sensor_msgs::msg::CameraInfo;
-
-  // Subscriptions
-  image_transport::SubscriberFilter sub_depth_, sub_intensity_;
-  message_filters::Subscriber<sensor_msgs::msg::CameraInfo> sub_info_;
-
-  int queue_size_;
-
-  // Publications
-  std::mutex connect_mutex_;
-  rclcpp::Publisher<PointCloud>::SharedPtr pub_point_cloud_;
-
-  using Synchronizer = message_filters::Synchronizer<SyncPolicy>;
-  std::shared_ptr<Synchronizer> sync_;
-
-  std::vector<double> D_;
-  std::array<double, 9> K_;
-
-  uint32_t width_;
-  uint32_t height_;
-
-  cv::Mat transform_;
-
-  void connectCb();
-
-  void imageCb(
-    const Image::ConstSharedPtr & depth_msg,
-    const Image::ConstSharedPtr & intensity_msg_in,
-    const CameraInfo::ConstSharedPtr & info_msg);
-
-  // Handles float or uint16 depths
-  template<typename T>
-  void convert_depth(
-    const Image::ConstSharedPtr & depth_msg,
-    PointCloud::SharedPtr & cloud_msg);
-
-  template<typename T>
-  void convert_intensity(
-    const Image::ConstSharedPtr & inten_msg,
-    PointCloud::SharedPtr & cloud_msg);
-
-  cv::Mat initMatrix(cv::Mat cameraMatrix, cv::Mat distCoeffs, int width, int height, bool radial);
-
-  rclcpp::Logger logger_ = rclcpp::get_logger("PointCloudXyziRadialNode");
-};
-
-cv::Mat PointCloudXyziRadialNode::initMatrix(
-  cv::Mat cameraMatrix, cv::Mat distCoeffs, int width,
-  int height, bool radial)
-{
-  int i, j;
-  int totalsize = width * height;
-  cv::Mat pixelVectors(1, totalsize, CV_32FC3);
-  cv::Mat dst(1, totalsize, CV_32FC3);
-
-  cv::Mat sensorPoints(cv::Size(height, width), CV_32FC2);
-  cv::Mat undistortedSensorPoints(1, totalsize, CV_32FC2);
-
-  std::vector<cv::Mat> ch;
-  for (j = 0; j < height; j++) {
-    for (i = 0; i < width; i++) {
-      cv::Vec2f & p = sensorPoints.at<cv::Vec2f>(i, j);
-      p[0] = i;
-      p[1] = j;
-    }
-  }
-
-  sensorPoints = sensorPoints.reshape(2, 1);
-
-  cv::undistortPoints(sensorPoints, undistortedSensorPoints, cameraMatrix, distCoeffs);
-
-  ch.push_back(undistortedSensorPoints);
-  ch.push_back(cv::Mat::ones(1, totalsize, CV_32FC1));
-  cv::merge(ch, pixelVectors);
-
-  if (radial) {
-    for (i = 0; i < totalsize; i++) {
-      normalize(
-        pixelVectors.at<cv::Vec3f>(i),
-        dst.at<cv::Vec3f>(i));
-    }
-    pixelVectors = dst;
-  }
-  return pixelVectors.reshape(3, width);
-}
 
 
 PointCloudXyziRadialNode::PointCloudXyziRadialNode(const rclcpp::NodeOptions & options)
@@ -163,7 +64,13 @@ PointCloudXyziRadialNode::PointCloudXyziRadialNode(const rclcpp::NodeOptions & o
     sub_depth_,
     sub_intensity_,
     sub_info_);
-  sync_->registerCallback(std::bind(&PointCloudXyziRadialNode::imageCb, this, _1, _2, _3));
+  sync_->registerCallback(
+    std::bind(
+      &PointCloudXyziRadialNode::imageCb,
+      this,
+      std::placeholders::_1,
+      std::placeholders::_2,
+      std::placeholders::_3));
 
   // Monitor whether anyone is subscribed to the output
   // TODO(ros2) Implement when SubscriberStatusCallback is available
@@ -236,76 +143,29 @@ void PointCloudXyziRadialNode::imageCb(
     transform_ = initMatrix(cv::Mat_<double>(3, 3, &K_[0]), cv::Mat(D_), width_, height_, true);
   }
 
+  // Convert Depth Image to Pointcloud
   if (depth_msg->encoding == enc::TYPE_16UC1) {
-    convert_depth<uint16_t>(depth_msg, cloud_msg);
+    convertDepthRadial<uint16_t>(depth_msg, cloud_msg, transform_);
   } else if (depth_msg->encoding == enc::TYPE_32FC1) {
-    convert_depth<float>(depth_msg, cloud_msg);
+    convertDepthRadial<float>(depth_msg, cloud_msg, transform_);
   } else {
     RCLCPP_ERROR(logger_, "Depth image has unsupported encoding [%s]", depth_msg->encoding.c_str());
     return;
   }
 
-  if (intensity_msg->encoding == enc::TYPE_16UC1) {
-    convert_intensity<uint16_t>(intensity_msg, cloud_msg);
-
-  } else if (intensity_msg->encoding == enc::MONO8) {
-    convert_intensity<uint8_t>(intensity_msg, cloud_msg);
+  if (intensity_msg->encoding == enc::MONO8) {
+    convertIntensity<uint8_t>(intensity_msg, cloud_msg);
+  } else if (intensity_msg->encoding == enc::MONO16) {
+    convertIntensity<uint16_t>(intensity_msg, cloud_msg);
+  } else if (intensity_msg->encoding == enc::TYPE_16UC1) {
+    convertIntensity<uint16_t>(intensity_msg, cloud_msg);
   } else {
     RCLCPP_ERROR(
-      logger_, "Intensity image has unsupported encoding [%s]",
-      intensity_msg->encoding.c_str());
+      logger_, "Intensity image has unsupported encoding [%s]", intensity_msg->encoding.c_str());
     return;
   }
 
   pub_point_cloud_->publish(*cloud_msg);
-}
-
-template<typename T>
-void PointCloudXyziRadialNode::convert_depth(
-  const Image::ConstSharedPtr & depth_msg,
-  PointCloud::SharedPtr & cloud_msg)
-{
-  // Combine unit conversion (if necessary) with scaling by focal length for computing (X,Y)
-  float bad_point = std::numeric_limits<float>::quiet_NaN();
-
-  sensor_msgs::PointCloud2Iterator<float> iter_x(*cloud_msg, "x");
-  sensor_msgs::PointCloud2Iterator<float> iter_y(*cloud_msg, "y");
-  sensor_msgs::PointCloud2Iterator<float> iter_z(*cloud_msg, "z");
-  const T * depth_row = reinterpret_cast<const T *>(&depth_msg->data[0]);
-
-  int row_step = depth_msg->step / sizeof(T);
-  for (int v = 0; v < static_cast<int>(cloud_msg->height); ++v, depth_row += row_step) {
-    for (int u = 0; u < static_cast<int>(cloud_msg->width); ++u, ++iter_x, ++iter_y, ++iter_z) {
-      T depth = depth_row[u];
-
-      // Missing points denoted by NaNs
-      if (!DepthTraits<T>::valid(depth)) {
-        *iter_x = *iter_y = *iter_z = bad_point;
-        continue;
-      }
-      const cv::Vec3f & cvPoint = transform_.at<cv::Vec3f>(u, v) * DepthTraits<T>::toMeters(depth);
-      // Fill in XYZ
-      *iter_x = cvPoint(0);
-      *iter_y = cvPoint(1);
-      *iter_z = cvPoint(2);
-    }
-  }
-}
-
-template<typename T>
-void PointCloudXyziRadialNode::convert_intensity(
-  const Image::ConstSharedPtr & intensity_msg,
-  PointCloud::SharedPtr & cloud_msg)
-{
-  sensor_msgs::PointCloud2Iterator<float> iter_i(*cloud_msg, "intensity");
-  const T * inten_row = reinterpret_cast<const T *>(&intensity_msg->data[0]);
-
-  const int i_row_step = intensity_msg->step / sizeof(T);
-  for (int v = 0; v < static_cast<int>(cloud_msg->height); ++v, inten_row += i_row_step) {
-    for (int u = 0; u < static_cast<int>(cloud_msg->width); ++u, ++iter_i) {
-      *iter_i = inten_row[u];
-    }
-  }
 }
 
 }  // namespace depth_image_proc
